orthopurify-code/requirements/requirements_llava.txt
absl-py==2.4.0
accelerate==0.30.1
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
bchlib==2.1.3
bitsandbytes==0.45.5
blessed==1.25.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
datasets==2.21.0
debugpy==1.8.20
decorator==5.2.1
deepspeed==0.14.4
dill==0.3.8
distlib==0.4.0
einops==0.6.1
einops-exts==0.0.4
et_xmlfile==2.0.0
evaluate==0.4.6
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.121.2
ffmpy==1.0.0
filelock==3.28.0
flatbuffers==25.12.19
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2024.6.1
ftfy==6.3.1
gast==0.7.0
google-auth==2.49.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpustat==1.1.1
gradio==4.16.0
gradio_client==0.8.1
grpcio==1.78.0
h11==0.14.0
h5py==3.16.0
hf-xet==1.4.3
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface_hub==0.36.2
idna==3.11
importlib_resources==6.5.2
ipykernel==7.2.0
ipython==8.38.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kagglehub==1.0.0
kagglesdk==0.1.18
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.9
latex2mathml==3.78.1
libclang==18.1.1
# LLaVA: 需要先 git clone 再 editable 安装，见下方注释
# git clone https://github.com/haotian-liu/LLaVA.git && cd LLaVA && pip install -e .
# 或直接：
llava @ git+https://github.com/haotian-liu/LLaVA.git
Markdown==3.10.2
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
modelscope==1.35.4
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.16
narwhals==2.11.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.13.0
nltk==3.9.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
openpyxl==3.1.5
opt_einsum==3.4.0
orjson==3.11.4
packaging==26.1
pandas==2.3.3
parso==0.8.6
peft==0.11.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==3.20.3
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycocoevalcap==1.2
pycocotools==2.0.11
pycparser==3.0
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pynvml==13.0.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-discovery==1.2.2
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
rarfile==4.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rouge-score==0.1.2
rpds-py==0.28.0
ruff==0.14.5
safetensors==0.6.2
scikit-learn==1.2.2
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.49.3
svgwrite==1.4.3
sympy==1.14.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.3.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.6.0
timm==0.6.13
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.1.2
torchaudio==2.11.0
torchvision==0.16.2
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.40.2
triton==2.1.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unrar==0.4
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==21.2.4
wavedrom==2.0.3.post3
wcwidth==0.2.14
websockets==11.0.3
Werkzeug==3.1.7
wrapt==2.1.2
xxhash==3.6.0
yarl==1.22.0
